requirements.txt
torch
torchvision
torchaudio
transformers
diffusers
accelerate 
num2words
openai
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-llms-replicate
llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-llms-huggingface-api
hf_xet
pyngrok
streamlit